Add fetch caching and remove only.util.injectHTML
Add fetch caching. only.utils.fetch and only.utils.fetchAndFill previously re-fetched from the provided URL at every invokation. This is not always desired, especially when server bandwidth is limited. This commit introduces the useCache parameter to both functions. When set to true, the response from the first fetch of a URL will be re-used. To force a new fetch (useful when the remote resource has been updated, for example), set useCache to false.

Remove only.util.injectHTML. It only performs node.innerHTML=text, which clearly can be done in a single line. It would be more useful as an alternative to only.fetch.fetchAndFill: only.fetch.fetchAndAppend. This may be added in a future update.