Replace most var declarations with let declarations
Went through and replaced most every var declaration with a let declaration. This was done for two reasons. First, it serves as a sanity check. Names cannot be let more than once within the same scope, so there is less risk of mistakenly redefining variables. Second, it removes at least one piece of weird code wherein an anonymous function is created only to copy a var in a loop. This is avoided by using let within the loop body as let scopes the to the block level, not the function level.

The only var declarations left alone were those used as loop indexers. This is because for loops with let declarations do not have well-defined behavior across browsers before ECMAScript 2015.