Make only.load return a Promise
only.load tells only.js to load a path. This loading occurs asynchronously via the use of Promises. In general use, it may be desirable to perform an action only after a page is loaded. To facilitate this, the only.load function now returns a promise that is guaranteed to resolve once the sitemap has been traversed and all directives have been invoked.