Add only.shouldUseSearchURLFormatting to only.js
Most webservers will 404 when they realize a path does not exist. only.js was assuming that the webserver was modified to ensure `index.html` is always loaded. However, this breaks the code for anyone giving it a try or for those who can't be bothered or are unable to change the behavior of their host/server.

This, alongside the fact that using `/?/` in the URL will always work (even after tweaking the webserver, motivated changing the default behavior of only.js to use the `/?/` approach.